Download product with multiple workers
This patch introduces workers for downloading files using the executable.
For programs using the library the parallelization should be done in the calling code itself.

There could be a lot of md5sums to process when working on long download periods.
Therefore `multiprocessing` is used to be able to calculate those hashes on multiple cores.